refactor: extract login page layout
